[N/A] Move types into internal
